MS Teams/requirements.txt
# Microsoft Bot Framework
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botbuilder-integration-aiohttp==4.17.0

# Web Framework
flask==3.0.0
aiohttp>=3.10,<4.0  # Required by botbuilder-integration-aiohttp 4.17.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# HTTP Client
httpx==0.25.2

# Redis (optional, for distributed rate limiting)
redis==5.0.8  # Includes async support, aioredis is deprecated and merged

# JSON and Data Handling
jsonschema==4.20.0
python-dateutil==2.8.2

# Security
cryptography==41.0.7
PyJWT==2.8.0

# Logging and Monitoring
python-json-logger==2.0.7
azure-monitor-opentelemetry==1.1.1  # Optional, for Application Insights

# Environment Variables
python-dotenv==1.0.0

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Type Hints
types-redis==4.6.0.11
types-aiofiles==23.2.0.0